Part 2: Adding Employers ViewModels:  ✅ A. Create a new ViewModel called AddEmployerViewModel that has 2 properties: Name and Location.
B.Add validation to both properties in the ViewModel so that both properties are required

Controllers: ✅
A. Set up a private DbContext variable so you can perform CRUD operations on the database. Pass it into a EmployerController constructor.

B. Complete Index() so that it passes all of the Employer objects in the database to the view.

C. Create an instance of AddEmployerViewModel inside of the Add() method and pass the instance into the View() return method.

D. Add the appropriate code to ProcessAddEmployerForm() so that it will process form submissions and make sure that only valid Employer objects are being saved to the database.

E. About() currently returns a view with vital
information about each employer such as their name and location. Make sure that the method is actually passing an Employer object to the view for display.
Views